Do not attempt to decode body for HEAD requests
Responses to HEAD requests are allowed to return Content-Type header for the
media type that would have been sent had the request been a GET.
However, we need not attempt to decode body because it is expected to not exist.

This fixes an error received when JSON.parse(null) is called because :body is
not in response.